@@ -3,8 +3,6 @@ apply plugin: "androidx.benchmark"
 apply plugin: 'kotlin-android'
 
 android {
-    compileSdkVersion 29
-
     compileOptions {
         sourceCompatibility = 1.8
         targetCompatibility = 1.8
@@ -16,9 +14,11 @@ android {
 
     defaultConfig {
         minSdkVersion 21
-        targetSdkVersion 29
+        targetSdkVersion 34
+        compileSdk 34
         versionCode 1
         versionName "1.0"
+        namespace "dev.andrewbailey.diff"
 
         testInstrumentationRunner 'androidx.benchmark.junit4.AndroidBenchmarkRunner'
         testInstrumentationRunnerArgument 'androidx.benchmark.suppressErrors', 'UNLOCKED'
@@ -56,8 +56,8 @@ dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation project(':difference')
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
-    testImplementation 'junit:junit:4.13'
-    androidTestImplementation 'androidx.test:runner:1.2.0'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.1'
-    androidTestImplementation "androidx.benchmark:benchmark-junit4:1.0.0"
+    testImplementation 'junit:junit:4.13.2'
+    androidTestImplementation 'androidx.test:runner:1.6.2'
+    androidTestImplementation 'androidx.test.ext:junit:1.2.1'
+    androidTestImplementation "androidx.benchmark:benchmark-junit4:1.3.3"
 }
