@@ -1,65 +1,42 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-    ext.kotlin_version = '1.3.72'
+    ext.kotlin_version = '2.0.21'
     repositories {
         google()
-        jcenter()
+        mavenCentral()
         maven { url "https://plugins.gradle.org/m2/" }
     }
 
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.6.3'
+        classpath 'com.android.tools.build:gradle:8.8.0'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-        classpath 'androidx.benchmark:benchmark-gradle-plugin:1.0.0'
-        classpath "org.jlleitschuh.gradle:ktlint-gradle:9.1.1"
-        classpath "org.jetbrains.dokka:dokka-gradle-plugin:0.10.1"
+        classpath 'androidx.benchmark:benchmark-gradle-plugin:1.3.3'
+        classpath "org.jlleitschuh.gradle:ktlint-gradle:12.1.2"
+        classpath "org.jetbrains.dokka:org.jetbrains.dokka.gradle.plugin:2.0.0"
     }
 }
 
 allprojects {
     apply plugin: "org.jlleitschuh.gradle.ktlint"
-    apply plugin: 'org.jetbrains.dokka'
 
     group = GROUP
     version = VERSION_NAME
 
     repositories {
         google()
-        jcenter()
+        mavenCentral()
     }
 
     ktlint {
+        version = "1.5.0"
         android = true
         verbose = true
-        disabledRules = [
-            "no-empty-class-body",
-            "no-blank-line-before-rbrace",
-            "no-wildcard-imports",
+        additionalEditorconfig = [
+           "max_line_length": "100"
         ]
-    }
+        disabledRules = [
 
-    dokka {
-        outputDirectory = "$rootDir/docs/1.x"
-        outputFormat = 'html'
-        multiplatform {
-            global {
-                perPackageOption {
-                    prefix = "dev.andrewbailey.difference"
-                    suppress = true
-                }
-            }
-            jvm {}
-        }
-    }
-
-    task dokkaJavadoc(type: org.jetbrains.dokka.gradle.DokkaTask) {
-        outputFormat = 'html'
-        outputDirectory = "$buildDir/javadoc"
-    }
-
-    task javadocJar(type: Jar, dependsOn: dokkaJavadoc) {
-        classifier = 'javadoc'
-        from "$buildDir/javadoc"
+        ]
     }
 }
