@@ -1,13 +1,15 @@
 package dev.andrewbailey.diff.impl
 
-import dev.andrewbailey.diff.impl.MyersDiffOperation.*
+import dev.andrewbailey.diff.impl.MyersDiffOperation.Delete
+import dev.andrewbailey.diff.impl.MyersDiffOperation.Insert
+import dev.andrewbailey.diff.impl.MyersDiffOperation.Skip
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
 class MyersDiffAlgorithmTest {
 
     @Test
-    fun `generateDiff with empty inputs returns empty sequence`() {
+    fun generateDiff_withEmptyInputs_returnsEmptySequence() {
         assertEquals(
             expected = emptyList(),
             actual = MyersDiffAlgorithm<Nothing>(
@@ -18,7 +20,7 @@ class MyersDiffAlgorithmTest {
     }
 
     @Test
-    fun `generateDiff with empty start returns additions`() {
+    fun generateDiff_withEmptyStart_returnsAdditions() {
         assertEquals(
             expected = listOf(Insert("A"), Insert("B"), Insert("C")),
             actual = MyersDiffAlgorithm(
@@ -29,7 +31,7 @@ class MyersDiffAlgorithmTest {
     }
 
     @Test
-    fun `generateDiff with empty end returns deletions`() {
+    fun generateDiff_withEmptyEnd_returnsDeletions() {
         assertEquals(
             expected = listOf(Delete, Delete, Delete),
             actual = MyersDiffAlgorithm(
@@ -40,7 +42,7 @@ class MyersDiffAlgorithmTest {
     }
 
     @Test
-    fun `generateDiff with same start and end returns skips`() {
+    fun generateDiff_withSameStartAndEnd_returnsSkips() {
         assertEquals(
             expected = listOf(Skip, Skip, Skip, Skip),
             actual = MyersDiffAlgorithm(
@@ -51,7 +53,7 @@ class MyersDiffAlgorithmTest {
     }
 
     @Test
-    fun `generateDiff with simple example`() {
+    fun generateDiff_withSimpleExample() {
         val original = "ABCABAC".toList()
         val updated = "CBABAC".toList()
 
@@ -73,7 +75,7 @@ class MyersDiffAlgorithmTest {
     }
 
     @Test
-    fun `generateDiff with DNA example`() {
+    fun generateDiff_withDnaExample() {
         val original = "tgtcgctctcaagatggcgtcttattacgaaaggagccagtccgggttgc".toList()
         val updated = "ggctggggttttcgcacggcgctccctccgcggttgtatctcaggcgaca".toList()
 
@@ -93,22 +95,26 @@ class MyersDiffAlgorithmTest {
     }
 
     @Test
-    fun `generateDiff with lorem ipsum example`() {
-        val original = ("Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod " +
-            "tempor incididunt ut labore et dolore magna aliqua. Quis auctor elit sed vulputate " +
-            "mi sit amet mauris commodo. Nec dui nunc mattis enim ut tellus elementum. Ultricies " +
-            "integer quis auctor elit sed vulputate mi sit amet. Ullamcorper velit sed " +
-            "ullamcorper morbi tincidunt.").toLowerCase().split(" ")
-        val updated = ("Malesuada fames ac turpis egestas. Varius sit amet mattis vulputate " +
-            "enim. Nisl nisi scelerisque eu ultrices vitae auctor eu augue. Sit amet volutpat " +
-            "consequat mauris nunc congue nisi vitae. Egestas purus viverra accumsan in nisl " +
-            "nisi scelerisque eu. Lobortis elementum nibh tellus molestie. Nulla at volutpat " +
-            "diam ut venenatis tellus in metus. Ac turpis egestas sed tempus urna et pharetra " +
-            "pharetra massa. Etiam sit amet nisl purus in mollis. Vivamus arcu felis bibendum ut " +
-            "tristique et egestas quis. Vestibulum lorem sed risus ultricies tristique nulla " +
-            "aliquet. Nunc scelerisque viverra mauris in aliquam. Facilisis magna etiam tempor " +
-            "orci eu lobortis elementum nibh. Purus faucibus ornare suspendisse sed nisi. Dui " +
-            "accumsan sit amet nulla.").toLowerCase().split(" ")
+    fun generateDiff_withLoremIpsumExample() {
+        val original = (
+            "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod " +
+                "tempor incididunt ut labore et dolore magna aliqua. Quis auctor elit sed " +
+                "vulputate mi sit amet mauris commodo. Nec dui nunc mattis enim ut tellus " +
+                "elementum. Ultricies integer quis auctor elit sed vulputate mi sit amet. " +
+                "Ullamcorper velit sed ullamcorper morbi tincidunt."
+            ).lowercase().split(" ")
+        val updated = (
+            "Malesuada fames ac turpis egestas. Varius sit amet mattis vulputate enim. Nisl " +
+                "nisi scelerisque eu ultrices vitae auctor eu augue. Sit amet volutpat consequat " +
+                "mauris nunc congue nisi vitae. Egestas purus viverra accumsan in nisl nisi " +
+                "scelerisque eu. Lobortis elementum nibh tellus molestie. Nulla at volutpat diam " +
+                "ut venenatis tellus in metus. Ac turpis egestas sed tempus urna et pharetra " +
+                "pharetra massa. Etiam sit amet nisl purus in mollis. Vivamus arcu felis " +
+                "bibendum ut tristique et egestas quis. Vestibulum lorem sed risus ultricies " +
+                "tristique nulla aliquet. Nunc scelerisque viverra mauris in aliquam. Facilisis " +
+                "magna etiam tempor orci eu lobortis elementum nibh. Purus faucibus ornare " +
+                "suspendisse sed nisi. Dui accumsan sit amet nulla."
+            ).lowercase().split(" ")
 
         val diff = MyersDiffAlgorithm(original, updated).generateDiff()
 
@@ -125,8 +131,8 @@ class MyersDiffAlgorithmTest {
         )
     }
 
-    private fun <T> applyDiff(original: List<T>, diff: Sequence<MyersDiffOperation<T>>): List<T> {
-        return original.toMutableList().apply {
+    private fun <T> applyDiff(original: List<T>, diff: Sequence<MyersDiffOperation<T>>): List<T> =
+        original.toMutableList().apply {
             var index = 0
             diff.forEach { operation ->
                 when (operation) {
@@ -136,6 +142,4 @@ class MyersDiffAlgorithmTest {
                 }
             }
         }
-    }
-
 }
