@@ -2,12 +2,9 @@ package dev.andrewbailey.diff.impl
 
 internal sealed class MyersDiffOperation<out T> {
 
-    data class Insert<T>(
-        val value: T
-    ) : MyersDiffOperation<T>()
+    data class Insert<T>(val value: T) : MyersDiffOperation<T>()
 
-    object Delete : MyersDiffOperation<Nothing>()
+    data object Delete : MyersDiffOperation<Nothing>()
 
-    object Skip : MyersDiffOperation<Nothing>()
-
+    data object Skip : MyersDiffOperation<Nothing>()
 }
